requirements.txt
defusedxml==0.7.1
et_xmlfile==2.0.0
numpy==2.3.2
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
